# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: F1 Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /predictions/winner:
    post:
      operationId: predictRaceWinner
      tags: [predictions]
      summary: Predict race winner based on factors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PredictionInput"
      responses:
        "200":
          description: Prediction results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PredictionResult"
...
components:
  schemas:
...
    PredictionInput:
      type: object
      required: [circuitId, season, round]
      properties:
        circuitId:
          type: string
        season:
          type: string
        round:
          type: integer
        weatherCondition:
          type: string
          enum: [dry, wet, mixed]
        temperature:
          type: ["number", "null"]
        humidity:
          type: ["number", "null"]
        windSpeed:
          type: ["number", "null"]
        tyreCompounds:
          type: array
          items:
            type: string
        safetyCarProbability:
          type: ["number", "null"]
    DriverPrediction:
      type: object
      required: [driverId, driverCode, driverName, constructorId, constructorName, constructorColor, winProbability, podiumProbability, factors]
      properties:
        driverId:
          type: string
        driverCode:
          type: string
        driverName:
          type: string
        constructorId:
          type: string
        constructorName:
          type: string
        constructorColor:
          type: string
        winProbability:
          type: number
        podiumProbability:
          type: number
        predictedPosition:
          type: integer
        factors:
          $ref: "#/components/schemas/PredictionFactors"
    PredictionFactors:
      type: object
      required: [recentForm, circuitHistory, qualifyingPace, tyreManagement, weatherAdaptability, carPerformance]
      properties:
        recentForm:
          type: number
        circuitHistory:
          type: number
        qualifyingPace:
          type: number
        tyreManagement:
          type: number
        weatherAdaptability:
          type: number
        carPerformance:
          type: number
    PredictionResult:
      type: object
      required: [circuitId, raceName, weatherCondition, predictions, modelConfidence, keyInsights]
      properties:
        circuitId:
          type: string
        raceName:
          type: string
        weatherCondition:
          type: string
        predictions:
          type: array
          items:
            $ref: "#/components/schemas/DriverPrediction"
        modelConfidence:
          type: number
        keyInsights:
          type: array
          items:
            type: string
...